# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Asset Tracker API",
    "version": "2.0.0",
    "description": "Comprehensive asset management system API with multi-tenancy, RBAC, webhooks, and advanced features.",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session token from BetterAuth authentication"
      }
    },
...
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/reservations/{id}": {
      "put": {
        "tags": ["Reservations"],
        "summary": "Update reservation (approve/reject/cancel)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["approved", "rejected", "cancelled"]
                  },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reservation updated" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  }
}